+++ src/main/java/me/bopis/king/mixin/mixins/MixinGuiNewChat.java
@@ -1,8 +1,5 @@
 package me.bopis.king.mixin.mixins;
 
-import java.awt.Color;
-import java.util.List;
-
 import me.bopis.king.Bopis;
 import me.bopis.king.features.modules.client.ClickGui;
 import me.bopis.king.features.modules.misc.ChatModifier;
@@ -17,7 +14,10 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-@Mixin(value={GuiNewChat.class})
+import java.awt.*;
+import java.util.List;
+
+@Mixin(value = {GuiNewChat.class})
 public class MixinGuiNewChat
         extends Gui {
     @Shadow
@@ -25,28 +25,28 @@ public class MixinGuiNewChat
     public List<ChatLine> drawnChatLines;
     private ChatLine chatLine;
 
-    @Redirect(method={"drawChat"}, at=@At(value="INVOKE", target="Lnet/minecraft/client/gui/GuiNewChat;drawRect(IIIII)V"))
+    @Redirect(method = {"drawChat"}, at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/GuiNewChat;drawRect(IIIII)V"))
     private void drawRectHook(int left, int top, int right, int bottom, int color) {
-        Gui.drawRect((int)left, (int)top, (int)right, (int)bottom, (int)( ChatModifier.getInstance().isOn() && ChatModifier.getInstance().clean.getValue() != false ? 0 : color));
+        Gui.drawRect(left, top, right, bottom, ChatModifier.getInstance().isOn() && ChatModifier.getInstance().clean.getValue() != false ? 0 : color);
     }
 
-    @Redirect(method={"drawChat"}, at=@At(value="INVOKE", target="Lnet/minecraft/client/gui/FontRenderer;drawStringWithShadow(Ljava/lang/String;FFI)I"))
+    @Redirect(method = {"drawChat"}, at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/FontRenderer;drawStringWithShadow(Ljava/lang/String;FFI)I"))
     private int drawStringWithShadow(FontRenderer fontRenderer, String text, float x, float y, int color) {
         if (text.contains("\u00a7+")) {
             float colorSpeed = 101 - 20;
-            Bopis.textManager.drawRainbowString(text, x, y, Color.HSBtoRGB( ClickGui.getInstance().hue, 1.0f, 1.0f), 100.0f, true);
+            Bopis.textManager.drawRainbowString(text, x, y, Color.HSBtoRGB(ClickGui.getInstance().hue, 1.0f, 1.0f), 100.0f, true);
         } else {
             Minecraft.getMinecraft().fontRenderer.drawStringWithShadow(text, x, y, color);
         }
         return 0;
     }
 
-    @Redirect(method={"setChatLine"}, at=@At(value="INVOKE", target="Ljava/util/List;size()I", ordinal=0, remap=false))
+    @Redirect(method = {"setChatLine"}, at = @At(value = "INVOKE", target = "Ljava/util/List;size()I", ordinal = 0, remap = false))
     public int drawnChatLinesSize(List<ChatLine> list) {
         return ChatModifier.getInstance().isOn() && ChatModifier.getInstance().infinite.getValue() != false ? -2147483647 : list.size();
     }
 
-    @Redirect(method={"setChatLine"}, at=@At(value="INVOKE", target="Ljava/util/List;size()I", ordinal=2, remap=false))
+    @Redirect(method = {"setChatLine"}, at = @At(value = "INVOKE", target = "Ljava/util/List;size()I", ordinal = 2, remap = false))
     public int chatLinesSize(List<ChatLine> list) {
         return ChatModifier.getInstance().isOn() && ChatModifier.getInstance().infinite.getValue() != false ? -2147483647 : list.size();
     }
